requirements.txt
# Core dependencies for image translation pipeline
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# OCR - PaddleOCR for better text detection and rotation support
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Alternative OCR option (commented out, can be used instead)
# pytesseract>=0.3.10
# pdf2image>=1.16.0  # Only needed if using tesseract with PDFs

# OpenAI API for translation
openai>=1.3.0

# Stable Diffusion inpainting
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.24.0

# Utilities
python-dotenv>=1.0.0
matplotlib>=3.7.0  # For visualization and debugging

